Split gopool checkDeadConnection into Linux and non-Linux file.
The only difference is that Linux adds POLLRDHUP to the flags sent to
poll. I should of course have checked that the code worked on multiple
systems rather than just assuming.
